Added support for "expandable" fieldsets and "selectfieldset" tabs.
When a fieldset is defined as "expandable" (boolean property), the
section is collapsed by default and may be expanded with a click.
The code includes the JavaScript code to collapse/expand the section.
It also adds an "expandable" class to the generated section for
styling purpose (e.g. to add an arrow next to the fieldset legend
to provide a visual clue that the section can be expanded). See
example in test/expandable.html

"selectfieldset" is a new type of form sections that should be used
for alternatives that are more complex than "pick up an element in
the list". It's typically useful for forms that provide multiple
entry points such as "search by user name", "search by playlist",
"search by text", etc. The "selectfieldset" type is similar to a
"fieldset" in that the same properties are used ("legend" if needed
and "items"). It is similar to a "select" field in that it provides
an alternative between different choices. Visually, the type is
represented as "tabs". Input fields from hidden tabs are disabled
so that they do not appear in the submitted form. Items of a
"selectfieldset" form section must be "optionfieldset" items. See
example in test/tabs.html