Added support for arrays with "complex" structures.
JSON Form now supports arrays whose items can follow an arbitrarily
complex structure (as opposed to being restricted to simple types
such as "string").

The important point to get is that references to array elements
from the form layout section must include [], e.g.:
 "foo.bar[].baz"

See test/complexlists.html for examples.

Restrictions:

1/ Items in an array must all be of the same type. In other words,
the "items property in the data schema for the array element must
either be an object or an array with one element. For instance:
 {
   "type": "array",
   "items": {
     "type": "object",
     "properties": {
       "key1": {
         "title": "Key 1",
         "type": "string"
       },
       "key2": {
         "title": "Key 2",
         "type": "string"
       }
     }
   }
 }

2/ Arrays cannot be nested for the time being.

3/ Similarly, a checkboxes section cannot be nested with an array.