Form values from options.value used for complex arrays as well.
Values were not properly used to initialize the form whenever the
array was more complex than a list of string items. Thanks to this
update, this should work relatively well ("relatively" because
arrays remain a kind of hack in JSON form for the time being, a
little bit of reformatting would be needed to do it well).