Event handlers of array items correctly set. Infinite loop possibility fixed.
The event handlers were only set for new items (in other words, those
added with a click on "Add more"). They are now correctly set for all
fields displayed when the form is rendered.

The script entered an infinite loop when the form contained an array
whose layout directly referenced a key in the schema and when there
existed values for some field, e.g.:
 {
  "schema": {
   "myarray": {
    "type": "array",
    "items": [ { "type": "string" } ]
   }
  },
  "form": [ "myarray" ],
  "value": { "somekey": "somevalue" }
 }