Defaults lib exported as a module.
The default library exports itself as an AMD module, a node.js
module or as a global JSONForm.setDefaultValues method, in that
order.

The setDefaultValues now takes an additional 'includeOptionalValues'
parameter. When set, default values for keys that are not defined as
"required" are also set.

Empty values are used for required keys that do not define any
default value:
 [] for an array
 {} for an object
 0 for a number or an integer
 '' for a string
 false for a boolean
 {} for everything else