New types added: "advancedfieldset" and "authfieldset".
The "advancedfieldset" type generates an expandable section with
an "Advanced options" legend.

The "authfieldset" type generates an expandable section with an
"Authentication settings" legend.

Both are more specific types of "fieldset" fields, and could be
created with a regular "fieldset" field (with the right legend
and flag). They are useful for the Factory in particular since it
treats such sections specifically.