Default values only proposed when form's "value" property not set.
The library used default values whenever it did not find the key
in the "value" object that may be attached to the JSON Form object
to symbolize the results of a previously submitted form.

The problem with that approach is that fields left empty are often
not included in the result (typically the case with checkboxes).
If their default value is not empty, the default value is always
used when the form is rendered. In short, a boolean flag whose
default value is "true" always appears as a checked checkbox, no
matter its previous value.

The change simply takes for granted that default values should only
be used when the form has never been submitted before, in other
words when the "value" property is not set.

Note an empty object counts as an actual form submission result
(it means all fields were left empty on purpose).