Array support completed. "Tab array" support added.
Add item, move up, move down, delete item should now work as expected.

In arrays, tab legends may use the value of an input field defined in
the array subtree, provided that field gets flagged with 'valueAsLegend'
and provided the tab legend string includes "{{value}}". This allows
to flag tabs with a useful title.

The index of the child may still be used. A tab legend "{{idx}}. {{value}}"
could yield "1. toto" if "toto" is the value of the input field to use
in legend.

The update includes various fixes. Consistency improved across code.