Moved libraries to "deps" folder and re-structured "tests".
The idea is to leave a clean "lib" folder that contains the actual
JSON Form library, and be more explicit about required dependencies
(in "deps" folder) and optional dependencies (in "deps/opt" folder).

For tests, the idea is to move to one test per file in an easy-to-grasp
folder structure to easily automate unit tests (not done yet).