Label and value templating now supported.
All properties that appear as labels can now be set to template
strings. JSON Form uses the "data" property of the JSON Form object
to fetch the appropriate values.

Example:
{
 schema: {
  textfield: {
   title: '{{hello}} {{user.name}}',
   type: 'string',
   'default': '{{def}}',
   description: '{{desc}}'
  }
 },
 form: [
  {
   key: 'textfield',
   prepend: '{{prepend}}',
   append: '{{append}}'
  }
 ],
 data: {
  hello: 'Hi',
  user: { name: 'tidoust' },
  def: 'Meh',
  desc: 'Foo',
  prepend: 'I am feeling',
  append: 'today'
 }
}

Specific values {{idx}} and {{value}} are reserved and replaced by
the index of the element in an array and the value of the first
child field that sets the "valueInLegend" property.

This feature may be presented as internationalization support,
although it's probably easier to just update the form part instead
of using the convoluted format shown in the example. That could still
be useful when JSON is needed (as opposed to JavaScript) and when
the form is created by some third-party entity.