Fixed array re-ordering bug.
JSON Form correctly switched values upon reordering but the jQuery
UI Sortable plug-in that it uses still changed the order of the DOM
elements, whereas that's not what we want. I could not find a simple
way to prevent the DOM element switching, so the fix is basically to
move the elements back to their initial position afterwards.