Added special templating option {{value foo[].bar.baz}}
The template gets replaced upon form initialization by the initial
value of the field targeted by the given key. If the key references
an enumeration, the final value is the result of running the
titleMap to the field's value if such a titleMap is defined.

This lets form authors initialize a field with the value of a
second field to start with.

The jsonform-defaults library still needs to be updated accordingly
to apply templating rules to the default values.