ACE input field: not injecting value in ACE <div> section anymore.
Not sure why, but the initial value was both injected in the <div>
section linked to the ACE editor and set in code through a call to
setValue. The injection did not escape the HTML, which caused
troubles when the value contained HTML tags, especially <script>
tags that were interpreted.