Fixed value initialialization for transloadit field.
The bug was due to a recent commit that introduced a nuance between
"value", defined as the stringified value of the field and
"node.value" that returns the actual value. In the case of the
transloadit, the field needs to use "node.value" as it performs
the conversion to a string on its own.