"selectfieldset": fields in other tabs were not disabled on init.
The fields in non selected alternative are automatically disabled
not to pollute the final result and not to create false errors,
but this was only done when the user changed the value of the
selection field, and not when the form was first loaded. Now fixed.

(old bug, probably there since the introduction of selectfieldset)