Arrays: drag-and-drop switched items. Should have moved dragged item.
A drag-and-drop within arrays should move the item being dragged
to the drop position, but JSON Form actually switched the item with
the item at the drop position. That's a possibility but definitely
not what the user would expect from a drag-and-drop.

This commit fixes the bug, both for arrays and tabarrays.