Fixed truncate function to include path up to next depth level (#9)
The "getPreviousNumberOfItems" function uses "truncateToArrayDepth"
to shorten the path to a key to the right depth level and then
compute the number of items at the level in previously submitted
values.

That second function did not return all the properties up to the
next array level, meaning the value checked in previously submitted
values could well be an object and not an array. For instance,
given a path:
 "foo.bar[].baz.stuff.blah[].thing"
the function used to return "foo.bar[]" for a depth of 1. It now
returns:
 "foo.bar[].baz.stuff.blah"

This should fix issue #9