Added spaces in button's template. New basic field type "msg" defined.
Bootstrap styles do not define any margin for buttons by default,
relying on spaces between buttons to display a natural gap on screen.
The HTML template now adds spaces before and after the <button> tag
for consecutive buttons not to touch each other.

The new "msg" basic field type lets forms include an HTML message.
The HTML message does not appear as a regular field but as a form
wide message (as opposed to the "help" field type).