New selectfieldset "tabsType" option. New "disabled" property.
The "selectfieldset" uses a selection box by default to render the
different choices. It may now use Bootstrap "tabs" or "pills"
based on a tabsType option. Example added.

Also added basic support for "disabled" property that disables a
field (sounds useless but could be needed in some cases). Note that
the "disabled" property does not interact well with alternatives:
the "selectfieldset" type toggles the "disabled" attribute
automatically and does not preserve the former value.