New "questions" and "question" field types added.
The "questions" field type renders a series of questions and computes
the final value of a schema key based on the answers provided.

When the form is rendered, only the first question is displayed.
Depending on the choice and the action(s) defined for that question,
the following question is displayed, or the key is set to a specific
value, or the form is submitted.

The "questions" field type is essentially useful to guide the user
through a series of choices with a few questions. It allows to build
simple flows such as:

1. Do you want to watch sports?
 - Yes: go to question 2,
 - No: go to question 3
2. What are you up to?
 - Live sports: set key to "livesports" and submit form
 - Rerun of previous events: set key to "oldsports" and submit form
3. What about news, then?
 - Yes: set key to "news" and submit form
 - No: go to question 4
4. Other things I have in store for you:
 - Drama: set key to "drama" and submit form
 - Comedy: set key to "comedy" and submit form
 - Thriller: set key to "thriller" and submit form

The final goal is always to set one (and only one) key to a specific
value. Choices may be rendered as radio buttons or as real buttons.
See playground for a short example.