The 'selectfieldset' field type may now be linked to a schema key.
The 'selectfieldset' field type used to be restricted to a purely
visual effect. It may make sense to bind it to a schema key to
save the selected tab when the form is submitted. This is now
possible.

The commit also rolls back on previous changes made to the field
type. In particuliar, it is no longer possible to use "pills".
The functionality should rather be implemented using radio buttons
in the same vein as "radiobuttons". To be done later on if need
arises.

Values that the schema key may take upon form submission must be
declared in the "enum" property of the schema key. JSON Form uses
the index of the choice in the absence of such a property. The
label associated with each choice can be defined using the "legend"
property of the field's children as before, or using a "titleMap"
as for regular enumerations.