"hideMenu" option to hide the selection menu of a "selectfieldset"
While this feature may look weird at first, it may make sense to
hide the selection menu when one wants to impose a given choice to
an alternative for some users while others get to choose (based
on some permission system for instance).

The "hideMenu" flag on a selectfieldset field does just that.

One may also want to disable the selection menu while still
rendering it with the "disabled" flag.

These flags are advanced flags, probably useless in 99% of all
cases.