setDefaultValues: option to skip required fields with no default value
The "setDefaultValues" function was typically built to create an
object that can be used to submit the form. As such, it sets required
fields to reasonable empty strings, objects, numbers, etc. when the
schema does not define any default for these fields.

The function may also be used upon form submission to detect whether
the user actually entered something on top of the default values
proposed by the schema. In that case, there should not be any empty
value for required fields.

The function now accepts a fourth flag param to skip required fields
that do not define a default value.