New array items: former submitted values re-used if available
When re-creating an array item that previously existed, the code
now re-uses former submitted values to initialize the created
fields. That's a way to recover from accidently suppressing an
array item. That's also a way to fix a small inconsistency in the
code as fields that reference another field's value indeed look
at former submitted values whereas more regular fields did not.