Updated JSON Form to support Bootstrap v2.1.1+
Only change is to separate text message from wrapped input field when
"append" or "prepend" is used.

In other words, JSON Form used to generate:

<div class="control-group">
  <label class="control-label">Nickname</label>
  <div class="groups input-prepend">
    <span class="add-on">@</span>
    <input type="text" />
    <span class="help-inline">Your nickname</span>
  </div>
</div>

That does not work in more recent versions because the text falls
under a "font-size: 0" CSS property triggered by "input-prepend".
The library now generates:

<div class="control-group">
  <label class="control-label">Nickname</label>
  <div class="groups">
    <div class="input-prepend">
      <span class="add-on">@</span>
      <input type="text" />
    </div>
    <span class="help-inline">Your nickname</span>
  </div>
</div>

Note this creates an additional level to reach the input field when
"prepend" or "append" is used.