New "split" library that splits a JSON Form object into two forms
The side library exposes a "split" function that takes a list of
schema keys (which must be first level keys in the schema) and a
JSON Form object. The function returns an object with two forms:
1. the "included" form that contains all the keys given as
parameters as well as all keys bound to these keys (because they
belong to the same "fieldset", array or grouping structure)
2. the "excluded" form that contains the rest of the keys.

The library is probably not very useful. The only use case being
when the JSON Form is generated by a third party but you want to
ask the user about specific keys first and then later on on the
rest.