Tabs: set "tolerance" option to "pointer" for jQuery UI Sortable
jQuery UI Sortable has a bug that makes moving an item in a list
to the first position a bit of a problem when the "containment"
option is used:
 http://bugs.jqueryui.com/ticket/5772

It makes sense to constrain possible moves to the list for tabs.
The "tolerance" option set to "pointer" is a quick and easy
workaround while this issue is addressed. It does not have any
side effect as far as I can tell.