file-transloadit: fix for results from steps other than ":original"
The "file-transloadit" field allows one to use arbitrary Transloadit
templates that can run any number of steps. The code assumed that the
final file would always be generated by the ":original" step which
may not always be the case.

Note JSON Form expects the upload to generate one and only one file.
It will pick up the first image that appears in the results otherwise,
but there is no guarantee that this image is the right one.