Field size dropped from HTML template, new "inputSize" param (#30)
JSON Form produced input types that all had the class "input-xlarge",
whereas it should not make any assumption on the styles and dimensions
of the input fields. The "input-xlarge" class is no longer added to
input fields by default.

To achieve the same effect, define an "inputSize" property in the
description of the input in the "form" section. If you want to apply
the same sizing class to all input fields, set that property in a
global "params" property of the JSON Form object:

{
  "schema": {},
  "form": [],
  "params": {
    "inputSize": "input-xlarge"
  }
}

PS: for some reason, the "range" input field did not make it to
previous commit. This commit also introduces the "range" input (#27)