Added support for "required", "maxLength" and "readonly" (#32)
JSON Form now converts JSON Schema properties "required" and
"maxLength" into the appropriate attributes on the input field.

For the "required" property, JSON Form also adds a "jsonform-required"
class to the field container's "<div>" and a "jsonform-hasrequired"
class to the "<form>" tag. The former allows one to flag required
fields easily with a CSS property such as:

  .jsonform-required > label:after {
    content: ' *';
    color: red;
  }

The latter allows one to add a legend at the end of the form, e.g.
with a CSS property such as:

  form.jsonform-hasrequired:after {
    content: '* Required field';
    display: block;
    color: red;
    padding-top: 1em;
  }

JSON Form also supports the "readonly" property when it is set in the
definition of an input field in the "form" section of a JSON Form
object (as opposed to in the schema since "readonly" simply does not
exist in JSON Schema).