Updated URI format to only accept HTTP(s) URIs
Note this is a willful deviation from the JSON Schema spec which does
not restrict URIs to HTTP(s) URIs. In practice, most if not all URIs
requested in a form are HTTP(s) URIs, so the restriction makes sense
not to force JSON Form users to specify their own regular expression
to restrict the scheme.