Dropped "required" attribute for checkbox fields
In JSON Schema, defining a boolean schema key as "required" means
that the property must appear in the generated object. Whether it
is true or false does not matter.

In HTML5, the "required" attribute on a checkbox means "the checkbox
must be checked to proceed". That's not really the same meaning.

JSONForm does not add the "required" attribute to a checkbox anymore
to remain compliant with the definition of "required" in JSON Schema.
As a matter of fact, since falsy values are now set in the generated
object, the "required" constraint on a boolean schema key is always
enforced.