Non empty object when a nested empty number was submitted (#38)
Given an object defined with a "number" property, form submission
always generated the path that led to that "number" property in the
resulting object, even when the user did not enter any actual number
in the form.

Paths should only be created when there are actual non empty
properties to create.