No support for "$ref". Added friendly error in the meantime (#43)
JSONForm does not yet support schemas that use the $ref keyword and
threw a pretty confusing error message when it received a schema that
did. The "getSchemaKey" function now correctly detects the use of the
"$ref" keyword and reports a friendly error message with a pointer to
issue #54 that tracks this use case.