Array/Tabarray: possibility to remove all items (#33)
There was no way to suppress the last item of an array. That
restriction was more for visual reasons (it usually does not look
great to only see "+/-" buttons) than for deep technical reasons.

In practice, not being able to remove all items creates artefact,
especially if one of the fields is rendered as a select box: there is
no way for JSONForm to tell whether the user actually selected
something in the select box or did not want to submit anything.
JSONForm always included the selected choice in the select box in the
submitted object, meaning there was no proper way to reset an array.

JSONForm now allows used to remove the last item in an array.

Note that, when an array is first rendered, it is still rendered with
one empty element as that's supposedly clearer than to display only
the "+/-" buttons. We may add an option later on to override that
setting and start with no array items if needed.