Array/Tabarray rendering: follow previous values if available (#33)
JSONForm renders an item for arrays the first time they are rendered
as it is usually better from a user experience perspective. However,
if the form has already been submitted and if previous values do not
contain any item for a given array, the array must be rendered without
items. JSONForm still rendered an item in that case.