Disabled unchecked checkboxes were included in submitted object
A recent change (#41) sends unchecked checkboxes as falsy values
when the form is submitted (instead of not sending them at all).

However, that change does not take into account the fact that
checkboxes may be in a "disabled" state. A disabled checkbox must
not take part in the construction of the submitted object.

Issues typically rise when a checkbox appears in more than one tabs
of a "selectfieldset". Fields that are in inactive tabs are correctly
flagged as "disabled" but since the code did not check that flag, it
would override a possibly checked value with the falsy value of the
last occurrence of the checkbox in the tabs of the selectfieldset.
Best illustrated with the example JSON Form object below where the
"flag" checkbox appears for Choice 1 and Choice 3:

{
  "schema": {
    "choice": {
      "type": "string",
      "title": "Make a choice",
      "enum": [ "Choice 1", "Choice 2", "Choice 3" ]
    },
    "flag": {
      "type": "boolean",
      "title": "Check me up"
    }
  },
  "form": [
    {
      "type": "selectfieldset",
      "key": "choice",
      "items": [
        flag,
        { "type": "section" },
        flag
      ]
    }
  ]
}

If "Choice 1" was selected and the checkbox checked, "flag" was not
set in the submitted object.