Fixed split library to always include "selectfieldset" children
When a "selectfieldset" references a schema key (which defines an
enumeration), the number of children of the field must always match
the number of values in the enumeration.

The split function did not preserve the direct children of a
selectfieldset if they did not reference one of the keys to include.