Added support for "readOnly" property in schema section (#74)
JSON Form already supported the "readonly" property but only in the
"form" section of the JSON Form object because I had missed the
"readOnly" keyword in JSON Schema (the keyword is only defined in
the JSON Hyper-Schema spec for some reason)

JSON Form now supports "readOnly" (same case as in JSON Schema) and
"readonly" (for backward compatibility) both in the "schema" and in
the "form" section of a JSON Form object.

Note that the form section cannot reset the flag if it is set in the
schema section.