A generic "object" key (no "properties") now creates a textarea field
JSON Form used to create a "fieldset" when it encountered a schema
key whose type was "object", expecting it to contain a "properties"
property that listed its properties and thus the fields of the
"fieldset".

However, schemas may contain generic "object" keys that do not
specify any properties. JSON Form created an empty fieldset in that
case. It now creates a "textarea" input field that expects an
object serialized as JSON.

Notes:
- the input field could easily be turned in an "ace" field whose
mode is "json", provided the right libraries are loaded.
- beware, for the time being, the code does not raise an error when
the string cannot be parsed as JSON. The submitted code contains an
empty object in that case.