ACE field: dropped sync call to "ace.require" (#78)
The ACE field was using a synchronous call to "ace.require" to
initialize the mode but that call only succeeds provided the mode
has already been loaded.

Using "setMode" does the exact same thing if the mode has already
been loaded and attempts to load the mode in the background otherwise.