The minItems/maxItems properties of a nested array were ignored (#79)
If the form referenced a schema key that contained nested array keys,
the "minItems" property of these nested array keys was simply ignored.
Moreover if the first array level contained a "minItems" property,
it was incorrectly applied to the nested array keys as well. Same
thing for "maxItems".

See issue #79 on Github for a concrete example.

The schema key reference was not properly processed. Two cases arise:
1. the form element references a schema key. That schema key is the
one to look for minItems/maxItems properties.
2. the form element does not reference a schema key, but some of its
children do. The closest parent array of the referenced schema keys
are the ones that may define minItems/maxItems properties.