Fix a minor bug in the playground app, resulting in 404's in some cases.
For some reason, the browser keeps appending a trailing '/' at the end
of the URLs. Since the loadExample function in playground.js parses the
request parameter itself and uses it to load the JSON, a trailing '/' at
the end of the URL causes the function to generate a request for a JSON
named '.json'.

The PR adds a check for a trailing '/' and taking appropriate action if
there is a trailing slash.