Legend to show enum values
Updated the binding of changes when using a legend. Changes were not reflected when the text field was using an enum unless you used they keyboard to scroll through the different options. Below is a sample form that has a legend with an array and text enum. Verified this works with an existing legend that is tied to a standard text box.

{
	"schema": {
	    "coverages":
	    {
	        "type":"array",
	        "items":{
				"type": "object",
				"title":"Max Coverage",
				"properties": {
					"name":{
						"type":"string",
						"title":"Coverage Name",
						"required":true,
						"enum":["","NPPA","MDDO","Scribe"]
					},
					"hours": {
						"type": "number",
						"title": "Coverage (Hours)",
						"required":true,
						"step":1
					}
				}				
	        }
		}		
	},
	"form": [
		{
		  "type": "tabarray",
		  "items": {
			"type": "section",
			"legend":"{{value}}",
			"items": [
				{
				  "key": "coverages[].name",
				  "title": "Coverage Name",
				  "valueInLegend": true
				},
				"coverages[].hours",
				{
      "type": "submit",
      "title": "Submit"
    } 
			]
		  }
		}
	]
}