feat: Add navigation tabs
This commit is to add a new feature: Navigation tabs. Navigation tabs are setup at 2 levels. The first is a type "tabs" that has a unique id and an items array. Each item is a object of type "tab" that has a title, which will be displayed as the tab text, and then items. These items are all of the fields to be presented on this tab.

Test HTML and JS pages have also been created under the tabs-nav namespace. This namespace was choosen so that it would not conflict with other "tab" namespaces that are currently present. It also is informative as it signifies the purpose of the tabs to be navigation, while still keeping the namespace brief.